Skip the preact prerender step because something in prerender calls 'window'
I tried disabling the API calls for prerender but it still failed -- I
think the call to 'window' comes at require-time (maybe from axios?). I
didn't bother diving deep on this one; the site is so lightweight that I
don't think the prerender is that important. We can always turn it back
on later