Remove the data submodule so I can make the data repo private
There's stuff in the data repo that needs the correct license information on it; it's technically breaching license for me to have a public github repo with the scraped text in it. When I expose the content through the Scroll Bible website, I will ensure the correct license information is displayed. By making this repo private for now I can postpone that license work