Fix a bug where navigating to another word and then back breaks the app
The observed behaviour was that going feed > verse > word > other word >
Browser back button causes the rendering to fail and the app to stop
responding right. The cause was that on the re-render of the first word,
it is first rendered with some data missing and then re-rendered. The
first render failed (accessing property on undefined object) preventing
the second render from happening. Now that the undefined is coalesced,
the first render succeeds (with nothing in place of the missing data).